#3 split main.go into observer.go, util.go, teams.go
